feat(timetable): split event form into three wizard steps
The modal becomes a wizard shell (SlideOver + WizardStepper + footer)
around three new step components, per the Planung-Redesign step table:

- event-form/step-termin.tsx (268 lines): Titel, Typ/Kategorie (the
  visibility rule 'expanded && isSeriesFlow' is unchanged, only the
  render location moved), Raum, Datum, Zeiten incl. quick presets,
  Tagesnotiz/Wochennotiz (cut decision: notes are Termin attributes,
  step 1; save semantics untouched)
- event-form/step-wiederholung.tsx (261 lines): repeat mode
  (einmalig/woechentlich/A-B), Wochentage, Planungszeitraum,
  Split/Ende actions in the edit case, series hints
- event-form/step-personal-kinder.tsx (485 lines): Zielgruppe (moved
  here per spec; cohort takeover identical), Kinderzuordnung
  (MultiSelectField, bulk, cohort), Personal incl. requiredStaff,
  conflict panel, coverage panel

Wizard shell (timetable-event-modal.tsx, 1269 -> 586 lines):
- footer: Abbrechen / Zurueck / Weiter (type=button, size=md) /
  Speichern; saving is possible from EVERY step once the Termin
  fields are valid (single event saves after step 1; defaults:
  einmalig, no staff)
- 'Weiter' validates only the current step's fields by filtering the
  hook's unchanged validateForm result via a STEP_FIELDS mapping; the
  final save keeps the full validation; a failed save jumps to the
  first step that shows the offending error
- entries: quick create / '+ Neu' / instance edit / series edit start
  at step 1; convertInstance opens step 2
- scope dialog, #1875 lost-edits chain and series-delete confirm stay
  in the shell; probes stay live via the hook regardless of the
  visible step

use-event-form.ts gains only a lastValidationErrors ref (synchronous
mirror of the last validateForm result) and exports validateForm -
no behavioral change.

plan-design-guards.test.ts: bright-color allowlist follows the moved
JSX (modal 2, field.tsx 1, step-wiederholung 3, step-personal-kinder 8;
total unchanged at 14).

Existing modal tests are deliberately NOT touched in this commit; the
step-navigation test update is the next, separately reviewable commit.
pnpm run check passes; design guards 144/144 green.
