+++ frontend/src/components/timetable/event-form/step-personal-kinder.tsx
@@ -1,0 +1,485 @@
+"use client";
+
+import { Alert } from "~/components/ui/alert";
+import { Button } from "~/components/ui/button";
+import { Input } from "~/components/ui/input";
+import { Tabs, TabsList, TabsTrigger } from "~/components/ui/tabs";
+import { timetableSelectClass } from "../timetable-style";
+import { Field } from "./field";
+import type { EventFormState, PersonOption } from "./form-model";
+import { MultiSelectField } from "./multi-select-field";
+import type { GroupOption } from "./use-event-form";
+import type {
+  ConflictWarningItem,
+  ShiftCoverageWarningItem,
+  TargetGroupType,
+} from "~/lib/timetable-types";
+
+const FORM_SELECT_CLASS = timetableSelectClass;
+
+/** Zielgruppe (target group) tab options, issue #1838. */
+const TARGET_GROUP_OPTIONS: Array<{ value: TargetGroupType; label: string }> = [
+  { value: "none", label: "Keine" },
+  { value: "jahrgang", label: "Jahrgang" },
+  { value: "klasse", label: "Klasse" },
+  { value: "gruppe", label: "Gruppe" },
+  { value: "angebot", label: "Angebotsauswahl" },
+];
+
+export interface StepPersonalKinderProps {
+  form: EventFormState;
+  update: <K extends keyof EventFormState>(
+    key: K,
+    value: EventFormState[K],
+  ) => void;
+  changeTargetGroupType: (nextType: TargetGroupType) => void;
+  fieldErrors: Record<string, string>;
+  groups: GroupOption[];
+  students: PersonOption[];
+  staff: PersonOption[];
+  loadingRefs: boolean;
+  loadingStudents: boolean;
+  studentLoadError: string | null;
+  loadingStaff: boolean;
+  staffLoadError: string | null;
+  retryStudentLoad: () => Promise<void>;
+  retryStaffLoad: () => Promise<void>;
+  expanded: boolean;
+  isSeriesFlow: boolean;
+  gradeLevelMax: number | undefined;
+  targetGradeOptions: Array<{
+    value: string;
+    label: string;
+    disabled: boolean;
+  }>;
+  preservesGradeAboveTenantCap: boolean;
+  studentBulkOptions: Array<{
+    key: string;
+    label: string;
+    memberIds: string[];
+  }>;
+  targetClassOptions: string[];
+  targetClassDescriptionIDs: string;
+  targetCohort: { label: string | null; memberIds: string[] };
+  missingTargetCohortCount: number;
+  targetCohortButtonLabel: string;
+  addTargetCohort: () => void;
+  conflictWarnings: ConflictWarningItem[];
+  coverageWarnings: ShiftCoverageWarningItem[];
+  coverageWarningCount: number;
+  coverageCheckError: string | null;
+  requiredStaffTouched: React.RefObject<boolean>;
+  staffRosterTouched: React.RefObject<boolean>;
+}
+
+/**
+ * Wizard step 3 "Personal und Kinder": Zielgruppe, the staff and student
+ * rosters, the Personalbedarf override, plus the advisory conflict and
+ * Dienstplan-coverage panel. JSX moved verbatim from the previous single-page
+ * form; Personal still renders before Kinder (Streichliste 8).
+ */
+export function StepPersonalKinder({
+  form,
+  update,
+  changeTargetGroupType,
+  fieldErrors,
+  groups,
+  students,
+  staff,
+  loadingRefs,
+  loadingStudents,
+  studentLoadError,
+  loadingStaff,
+  staffLoadError,
+  retryStudentLoad,
+  retryStaffLoad,
+  expanded,
+  isSeriesFlow,
+  gradeLevelMax,
+  targetGradeOptions,
+  preservesGradeAboveTenantCap,
+  studentBulkOptions,
+  targetClassOptions,
+  targetClassDescriptionIDs,
+  targetCohort,
+  missingTargetCohortCount,
+  targetCohortButtonLabel,
+  addTargetCohort,
+  conflictWarnings,
+  coverageWarnings,
+  coverageWarningCount,
+  coverageCheckError,
+  requiredStaffTouched,
+  staffRosterTouched,
+}: Readonly<StepPersonalKinderProps>) {
+  let studentRosterField: React.ReactNode;
+  if (loadingStudents) {
+    studentRosterField = (
+      <Alert
+        type="info"
+        message="Kinderliste wird geladen … Die bestehende Kinderzuordnung bleibt beim Speichern unverändert."
+      />
+    );
+  } else if (studentLoadError) {
+    studentRosterField = (
+      <div className="flex flex-col gap-2">
+        <Alert type="warning" message={studentLoadError} />
+        <Button
+          type="button"
+          variant="outline"
+          size="compact"
+          className="self-start"
+          onClick={() => void retryStudentLoad()}
+        >
+          Kinder erneut laden
+        </Button>
+      </div>
+    );
+  } else {
+    studentRosterField = (
+      <MultiSelectField
+        label="Kinder"
+        options={students}
+        value={form.studentIds}
+        onChange={(ids) => update("studentIds", ids)}
+        metadata="student"
+        bulkOptions={studentBulkOptions}
+      />
+    );
+  }
+
+  let staffRosterField: React.ReactNode;
+  if (loadingStaff) {
+    staffRosterField = (
+      <Alert
+        type="info"
+        message="Personalliste wird geladen … Die bestehende Personalzuordnung bleibt beim Speichern unverändert."
+      />
+    );
+  } else if (staffLoadError) {
+    staffRosterField = (
+      <div className="flex flex-col gap-2">
+        <Alert type="warning" message={staffLoadError} />
+        <Button
+          type="button"
+          variant="outline"
+          size="compact"
+          className="self-start"
+          onClick={() => void retryStaffLoad()}
+        >
+          Personal erneut laden
+        </Button>
+      </div>
+    );
+  } else {
+    staffRosterField = (
+      <>
+        <MultiSelectField
+          label="Personal"
+          options={staff}
+          value={form.staffIds}
+          onChange={(ids) => {
+            staffRosterTouched.current = true;
+            update("staffIds", ids);
+            if (form.primaryStaffId && !ids.includes(form.primaryStaffId)) {
+              update("primaryStaffId", "");
+            }
+          }}
+          metadata="staff"
+        />
+
+        {isSeriesFlow && form.staffIds.length > 0 && (
+          <Field label="Zuständige Person" htmlFor="event_primary_staff">
+            <select
+              id="event_primary_staff"
+              value={form.primaryStaffId}
+              onChange={(event) => {
+                staffRosterTouched.current = true;
+                update("primaryStaffId", event.target.value);
+              }}
+              className={FORM_SELECT_CLASS}
+            >
+              <option value="">Keine Auswahl</option>
+              {staff
+                .filter((person) => form.staffIds.includes(person.id))
+                .map((person) => (
+                  <option key={person.id} value={person.id}>
+                    {person.name}
+                  </option>
+                ))}
+            </select>
+          </Field>
+        )}
+      </>
+    );
+  }
+
+  return (
+    <>
+      {expanded && isSeriesFlow && (
+        <div className="flex flex-col gap-1">
+          <span className="text-xs font-semibold text-gray-700">
+            Zielgruppe
+          </span>
+          <Tabs
+            value={form.targetGroupType}
+            onValueChange={(value) =>
+              changeTargetGroupType(value as TargetGroupType)
+            }
+          >
+            <TabsList aria-label="Zielgruppe" className="w-fit">
+              {TARGET_GROUP_OPTIONS.map((option) => (
+                <TabsTrigger key={option.value} value={option.value}>
+                  {option.label}
+                </TabsTrigger>
+              ))}
+            </TabsList>
+          </Tabs>
+
+          {form.targetGroupType === "jahrgang" && (
+            <div className="mt-1">
+              <Field
+                label="Jahrgang"
+                htmlFor="event_target_grade_level"
+                required
+                error={fieldErrors.targetGradeLevel}
+              >
+                <select
+                  id="event_target_grade_level"
+                  value={form.targetGradeLevel}
+                  onChange={(event) =>
+                    update("targetGradeLevel", event.target.value)
+                  }
+                  required
+                  aria-invalid={fieldErrors.targetGradeLevel ? true : undefined}
+                  aria-describedby={
+                    fieldErrors.targetGradeLevel
+                      ? "event_target_grade_level_error"
+                      : undefined
+                  }
+                  className={FORM_SELECT_CLASS}
+                >
+                  <option value="">Jahrgang wählen …</option>
+                  {targetGradeOptions.map((option) => (
+                    <option
+                      key={option.value}
+                      value={option.value}
+                      disabled={option.disabled}
+                    >
+                      {option.label}
+                    </option>
+                  ))}
+                </select>
+              </Field>
+              {preservesGradeAboveTenantCap ? (
+                <p className="mt-1 text-xs text-gray-500" role="status">
+                  Jahrgang {form.targetGradeLevel} liegt über der aktuell
+                  konfigurierten Höchststufe {gradeLevelMax}. Die bestehende
+                  Zielgruppe bleibt beim Speichern erhalten.
+                </p>
+              ) : null}
+            </div>
+          )}
+
+          {form.targetGroupType === "klasse" && (
+            <div className="mt-1">
+              <Field
+                label="Klasse"
+                htmlFor="event_target_school_class"
+                required
+                error={fieldErrors.targetSchoolClass}
+              >
+                <select
+                  id="event_target_school_class"
+                  value={form.targetSchoolClass}
+                  onChange={(event) =>
+                    update("targetSchoolClass", event.target.value)
+                  }
+                  disabled={loadingStudents || studentLoadError !== null}
+                  required
+                  aria-invalid={
+                    fieldErrors.targetSchoolClass ? true : undefined
+                  }
+                  aria-describedby={targetClassDescriptionIDs || undefined}
+                  className={FORM_SELECT_CLASS}
+                >
+                  <option value="">Klasse wählen …</option>
+                  {targetClassOptions.map((schoolClass) => (
+                    <option key={schoolClass} value={schoolClass}>
+                      {schoolClass}
+                    </option>
+                  ))}
+                </select>
+              </Field>
+              {loadingStudents || studentLoadError ? (
+                <p
+                  id="event_target_school_class_availability"
+                  className="mt-1 text-xs text-gray-500"
+                  role="status"
+                >
+                  {studentLoadError
+                    ? "Die Klassenliste ist nicht verfügbar. Eine bestehende Klassen-Zielgruppe bleibt unverändert."
+                    : "Klassenliste wird geladen …"}
+                </p>
+              ) : null}
+            </div>
+          )}
+
+          {form.targetGroupType === "gruppe" && (
+            <div className="mt-1">
+              <Field
+                label="Gruppe"
+                htmlFor="event_target_gruppe"
+                required
+                error={fieldErrors.educationGroupId}
+              >
+                <select
+                  id="event_target_gruppe"
+                  value={form.educationGroupId}
+                  onChange={(event) =>
+                    update("educationGroupId", event.target.value)
+                  }
+                  disabled={loadingRefs}
+                  required
+                  aria-invalid={fieldErrors.educationGroupId ? true : undefined}
+                  aria-describedby={
+                    fieldErrors.educationGroupId
+                      ? "event_target_gruppe_error"
+                      : undefined
+                  }
+                  className={FORM_SELECT_CLASS}
+                >
+                  <option value="">Gruppe wählen …</option>
+                  {groups.map((group) => (
+                    <option key={group.id} value={group.id}>
+                      {group.name}
+                    </option>
+                  ))}
+                </select>
+              </Field>
+            </div>
+          )}
+
+          {form.targetGroupType === "angebot" && (
+            <p className="mt-1 text-xs text-gray-500">
+              Kinder kommen automatisch über ein Betreuungsangebot hinzu. Die
+              Verknüpfung wird unter „Angebote“ beim jeweiligen Angebot gepflegt
+              (Feld „Regeltermin“).
+            </p>
+          )}
+
+          {targetCohort.label && !loadingStudents && !studentLoadError && (
+            <div className="mt-2 flex flex-wrap items-center justify-between gap-2 rounded-xl border border-gray-200 bg-white/70 p-3">
+              <p className="min-w-0 flex-1 text-xs leading-5 text-gray-600">
+                Die Zielgruppe beschreibt den Regeltermin. Übernimm die
+                passenden Kinder mit einem Klick; die Auswahl bleibt danach
+                anpassbar.
+              </p>
+              <Button
+                type="button"
+                variant="outline"
+                size="compact"
+                onClick={addTargetCohort}
+                disabled={
+                  targetCohort.memberIds.length === 0 ||
+                  missingTargetCohortCount === 0
+                }
+              >
+                {targetCohortButtonLabel}
+              </Button>
+            </div>
+          )}
+        </div>
+      )}
+
+      {staffRosterField}
+
+      <Field label="Benötigtes Personal" htmlFor="event_required_staff">
+        <Input
+          id="event_required_staff"
+          type="number"
+          min={0}
+          step={1}
+          inputMode="numeric"
+          value={form.requiredStaff}
+          onChange={(event) => {
+            requiredStaffTouched.current = true;
+            update("requiredStaff", event.target.value);
+          }}
+          placeholder="automatisch aus Betreuungsschlüssel"
+          controlSize="compact"
+        />
+        <p className="mt-1 text-xs text-gray-500">
+          Leer = automatisch: Es gilt der Wert der Terminreihe, sonst die
+          Berechnung aus dem Betreuungsschlüssel (Kinderzahl). Eine Zahl legt
+          den Bedarf fest und überschreibt beides.
+        </p>
+      </Field>
+
+      {studentRosterField}
+
+      {(conflictWarnings.length > 0 ||
+        coverageWarnings.length > 0 ||
+        coverageCheckError) && (
+        // Advisory pre-save hints (QA M7): never disables Speichern.
+        <div className="flex flex-col gap-2 border-t border-gray-200 pt-3">
+          <p className="sr-only" aria-live="polite">
+            {`${conflictWarnings.length} Terminüberschneidungen und ${coverageWarningCount} Dienstplan-Lücken gefunden. Speichern ist weiterhin möglich.`}
+          </p>
+          {conflictWarnings.map((warning, index) => (
+            <Alert
+              key={`${warning.kind}-${warning.resourceId}-${index}`}
+              type="warning"
+              message={`Hinweis: ${warning.message}`}
+              announce="off"
+            />
+          ))}
+          {coverageWarningCount > 0 && (
+            <Alert
+              type="warning"
+              message={`${coverageWarningCount} Dienstplan-${coverageWarningCount === 1 ? "Lücke" : "Lücken"} gefunden. Speichern ist weiterhin möglich.`}
+              announce="off"
+            />
+          )}
+          {coverageWarnings.slice(0, 3).map((warning, index) => (
+            <p
+              key={`shift-coverage-example-${warning.staffId}-${warning.date}-${warning.uncoveredStartTime}-${index}`}
+              className="rounded-lg border border-yellow-100 bg-yellow-50/50 px-3 py-2 text-sm text-yellow-800"
+            >
+              {warning.message}
+            </p>
+          ))}
+          {coverageWarningCount > 3 && (
+            <details className="rounded-lg border border-yellow-100 bg-yellow-50/40 px-3 py-2 text-sm text-yellow-800">
+              <summary className="cursor-pointer font-medium focus-visible:ring-2 focus-visible:ring-yellow-600 focus-visible:outline-none">
+                {coverageWarningCount - 3} weitere Lücken anzeigen
+              </summary>
+              <div className="mt-2 max-h-48 space-y-2 overflow-y-auto overscroll-contain pr-1">
+                {coverageWarnings.slice(3).map((warning, index) => (
+                  <p
+                    key={`shift-coverage-detail-${warning.staffId}-${warning.date}-${warning.uncoveredStartTime}-${index}`}
+                    className="border-t border-yellow-100 pt-2 first:border-t-0 first:pt-0"
+                  >
+                    {warning.message}
+                  </p>
+                ))}
+                {coverageWarningCount > coverageWarnings.length && (
+                  <p className="border-t border-yellow-100 pt-2 font-medium">
+                    Es werden höchstens 100 Beispiele angezeigt.
+                  </p>
+                )}
+              </div>
+            </details>
+          )}
+          {coverageCheckError && (
+            <Alert
+              type="warning"
+              message={`Hinweis: ${coverageCheckError}`}
+              announce="off"
+            />
+          )}
+        </div>
+      )}
+    </>
+  );
+}
